requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.6
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-modeltranslation==0.18.8
django-mptt==0.14.0
django-quill==0.1.0
django-quill-editor==0.1.40
django-recaptcha==3.0.0
django-rest-swagger==2.2.0
django-rq==2.4.1
django-tinymce==3.5.0
django-translation-manager==1.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.5.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
openapi-codec==1.3.2
Pillow==9.4.0
polib==1.2.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pyrsistent==0.19.3
python-gettext==5.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.5
requests==2.28.2
rq==1.14.1
simplejson==3.18.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
